requirements.txt
langgraph>=0.2.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
chromadb>=0.4.0
docker>=7.0.0
GitPython>=3.1.0
